+++ src/main/java/com/twoCube/gifts/controller/GiftController.java
@@ -1,22 +1,12 @@
 package com.twoCube.gifts.controller;
 
-import com.twoCube.common.annotation.CurrentUser;
-import com.twoCube.gifts.domain.GiftNote;
-import com.twoCube.gifts.dto.*;
+import com.twoCube.gifts.dto.request.MainRequest;
 import com.twoCube.gifts.service.*;
-import com.twoCube.members.domain.Member;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
-import springfox.documentation.annotations.ApiIgnore;
-
-import java.util.ArrayList;
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/v1/gifts")
