+++ src/main/java/com/twoCube/gifts/service/GiftFlowerService.java
@@ -1,6 +1,6 @@
 package com.twoCube.gifts.service;
 
-import com.twoCube.gifts.dto.FlowerRequest;
+import com.twoCube.gifts.dto.request.FlowerRequest;
 import com.twoCube.gifts.dto.FlowerResponse;
 import com.twoCube.members.domain.Member;
 import org.springframework.stereotype.Service;
