+++ src/main/java/com/twoCube/gifts/service/GiftNoteService.java
@@ -1,8 +1,7 @@
 package com.twoCube.gifts.service;
 
-import com.twoCube.gifts.dto.NoteRequest;
+import com.twoCube.gifts.dto.request.NoteRequest;
 import com.twoCube.members.domain.Member;
-import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
