+++ src/main/java/com/twoCube/gifts/service/GiftPillService.java
@@ -1,7 +1,7 @@
 package com.twoCube.gifts.service;
 
 import com.twoCube.gifts.dto.PillListResponse;
-import com.twoCube.gifts.dto.PillRequest;
+import com.twoCube.gifts.dto.request.PillRequest;
 import com.twoCube.members.domain.Member;
 import org.springframework.stereotype.Service;
